Remove or enable commented-out rules
Some of the rules were originally commented out because the UI wasn't in
a state where the rules were ready to be included. However, for the
shared ESLint file, I think most of these rules should be enabled. If we
absolutely need to turn the off for a specific project (such as the UI),
we can easily do so.